+++ joss-paper/paper.bib
@@ -18,7 +18,7 @@
     title   = {An open source device for operant licking in rats.},
     year    = {2017},
     volume  = {5},
-    doi     = {10.7717/peerj.2981.}
+    doi     = {10.7717/peerj.2981}
 }
 
 
