### 2024-11-16 9dd92e0
removed unnecessary SPI configuration, moved logic to FourDigitSevenSegmentDisplayHelper.java, made something that actually works as tested with hardware :)
