+++ components/src/main/resources/application.yml
@@ -10,11 +10,6 @@ pi4j:
       address: 8            # <3>
       baud: 500000          # <4>
       reset-pin: 25         # <5>
-    four-digit-seven-segment:
-      name: Four Digit Seven Segment Display
-      address: 0
-      baud: 1000000
-      reset-pin: 25
   # end::spi[]
 
   # tag::pwm[]
