### 2024-11-18 93bd895
Simplified decimal point, moved error handling, support more characters on display
