### 2024-11-18 973de47
Add setLog and getDisplayValue, add value validation, add doc comments, update method name
