+++ pi4micronaut-utils/src/main/java/com/opensourcewithslu/outputdevices/FourDigitSevenSegmentDisplayHelper.java
@@ -7,7 +7,7 @@ import org.slf4j.LoggerFactory;
 import java.util.Map;
 
 public class FourDigitSevenSegmentDisplayHelper {
-    private static final Logger log = LoggerFactory.getLogger(FourDigitSevenSegmentDisplayHelper.class);
+    private static Logger log = LoggerFactory.getLogger(FourDigitSevenSegmentDisplayHelper.class);
 
     private final DigitalOutput sdi;
     private final DigitalOutput rclk;
@@ -17,6 +17,7 @@ public class FourDigitSevenSegmentDisplayHelper {
     private final DigitalOutput digit2;
     private final DigitalOutput digit3;
 
+    private String displayValue;
     private boolean enabled = false;
     private final int decimalPoint = 0x7f;
 
@@ -46,9 +47,21 @@ public class FourDigitSevenSegmentDisplayHelper {
             Map.entry('F', 0x8e)
     );
 
-    private String displayValue;
-
-    public FourDigitSevenSegmentDisplayHelper(DigitalOutput sdi, DigitalOutput rclk, DigitalOutput srclk, DigitalOutput digit0, DigitalOutput digit1, DigitalOutput digit2, DigitalOutput digit3) {
+    /**
+     * Constructor for the FourDigitSevenSegmentDisplayHelper class.
+     *
+     * @param sdi    Serial data input
+     * @param rclk   Register clock
+     * @param srclk  Shift register clock
+     * @param digit0 The first digit
+     * @param digit1 The second digit
+     * @param digit2 The third digit
+     * @param digit3 The fourth digit
+     */
+    //tag::const[]
+    public FourDigitSevenSegmentDisplayHelper(DigitalOutput sdi, DigitalOutput rclk, DigitalOutput srclk, DigitalOutput digit0, DigitalOutput digit1, DigitalOutput digit2, DigitalOutput digit3)
+    //end::const[]
+    {
         this.sdi = sdi;
         this.rclk = rclk;
         this.srclk = srclk;
@@ -60,7 +73,7 @@ public class FourDigitSevenSegmentDisplayHelper {
         this.displayValue = "";
     }
 
-    public void shiftOut(Integer data, boolean decimalPointEnabled) {
+    private void shiftOut(Integer data, boolean decimalPointEnabled) {
         int value;
         for (int i = 0; i < 8; i++) { // Loop through each bit in the byte, one for each of the 7 segment and the decimal point
             if (decimalPointEnabled) {
@@ -80,7 +93,13 @@ public class FourDigitSevenSegmentDisplayHelper {
         rclk.low();
     }
 
-    public void clear() {
+    /**
+     * Clears the display.
+     */
+    //tag::method[]
+    public void clear()
+    //end::method[]
+    {
         displayValue = "";
     }
 
@@ -111,32 +130,83 @@ public class FourDigitSevenSegmentDisplayHelper {
             throw new IllegalArgumentException("Character is not supported by seven-segment display");
         }
 
-//        log.info("previous value: " + numbers[Character.getNumericValue(c)] + " and now trying to do " + value);
-
         try {
-//            shiftOut(numbers[Character.getNumericValue(c)], decimalPoint);
             shiftOut(value, decimalPoint);
         } catch (Exception e) {
             log.error(String.format("Error shifting out value %d to digit %d:", value, digit), e);
         }
     }
 
-    public void enable() {
+    /**
+     * Enables the display.
+     */
+    //tag::method[]
+    public void enable()
+    //end::method[]
+    {
         this.enabled = true;
         this.startDisplayThread();
     }
 
-    public void disable() {
+    /**
+     * Disables the display.
+     */
+    //tag::method[]
+    public void disable()
+    //end::method[]
+    {
[…]
