### 2024-11-18 a68bb7a
unnecessary application.yml config, fixed test, undid accidental change a previous commit had done to the servo motor config
