+++ components/src/main/java/com/opensourcewithslu/components/controllers/FourDigitSevenSegmentDisplayController.java
@@ -6,6 +6,7 @@ import io.micronaut.http.annotation.Controller;
 import io.micronaut.http.annotation.Get;
 import jakarta.inject.Named;
 
+//tag::ex[]
 @Controller("/four-digit-seven-segment")
 public class FourDigitSevenSegmentDisplayController {
     private final FourDigitSevenSegmentDisplayHelper displayHelper;
@@ -40,3 +41,4 @@ public class FourDigitSevenSegmentDisplayController {
         displayHelper.clear();
     }
 }
+//end::ex[]
