+++ components/src/main/resources/application.yml
@@ -35,7 +35,7 @@ pi4j:
       shutdown: 0
     servo-motor:
       name: Servo Motor
-      address: 17
+      address: 18
       pwmType: SOFTWARE
       provider: pigpio-pwm
       initial: 0
@@ -62,10 +62,6 @@ pi4j:
       name: lcd                           # <2>
       bus: 1                              # <3>
       device: 0x27                        # <4>
-    four-digit-seven-segment:
-      name: 4 Digit 7 Segment Display
-      bus: 1
-      device: 0x27
   # end::i2c[]
 
   # tag::digitalInput[]
