### 2024-11-18 0799a20
Revert "4 Digit 7 Segment Display"
