+++ pi4micronaut-utils/src/test/java/com/opensourcewithslu/outputdevices/FourDigitSevenSegmentDisplayHelperTest.java
@@ -1,266 +1,0 @@
-package com.opensourcewithslu.outputdevices;
-
-import com.pi4j.context.ContextProperties;
-import com.pi4j.io.gpio.digital.DigitalOutput;
-import com.pi4j.io.i2c.I2CConfig;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import org.slf4j.Logger;
-
-import static org.mockito.Mockito.*;
-import static org.junit.jupiter.api.Assertions.*;
-
-import com.pi4j.context.Context;
-import com.pi4j.crowpi.components.SevenSegmentComponent;
-
-import java.lang.reflect.Field;
-
-public class FourDigitSevenSegmentDisplayHelperTest {
-    DigitalOutput sdi = mock(DigitalOutput.class);
-    DigitalOutput rclk = mock(DigitalOutput.class);
-    DigitalOutput srclk = mock(DigitalOutput.class);
-    DigitalOutput digit0 = mock(DigitalOutput.class);
-    DigitalOutput digit1 = mock(DigitalOutput.class);
-    DigitalOutput digit2 = mock(DigitalOutput.class);
-    DigitalOutput digit3 = mock(DigitalOutput.class);
-    FourDigitSevenSegmentDisplayHelper displayHelper = new FourDigitSevenSegmentDisplayHelper(sdi, rclk, srclk, digit0, digit1, digit2, digit3);
-    Logger log = mock(Logger.class);
-
-    /*@BeforeEach
-    void setUp() throws Exception {
-        // Mock the Context to return a non-null ContextProperties
-        when(pi4jContext.properties()).thenReturn(contextProperties);
-
-        displayHelper = new FourDigitSevenSegmentDisplayHelper(i2cConfig, pi4jContext);
-
-        // Use reflection to set the mock SevenSegmentComponent
-        Field displayField = FourDigitSevenSegmentDisplayHelper.class.getDeclaredField("display");
-        displayField.setAccessible(true);
-        displayField.set(displayHelper, displayComponent);
-    }*/
-
-    @BeforeEach
-    public void openMocks() {
-        displayHelper.setLog(log);
-    }
-
-    @Test
-    void longNumberFails() {
-        String value = "12345";
-        displayHelper.displayValue(value);
-        verify(log).error("Display value must not have more than 4 non-decimal point characters");
-        verify(log, never()).info("Displaying value: {}", "12345");
-    }
-
-    @Test
-    void consecutiveDecimalPointsFails() {
-        String value = "1..23";
-        displayHelper.displayValue(value);
-        verify(log).error("Display value cannot have consecutive decimal points");
-        verify(log, never()).info("Displaying value: {}", "1..23");
-    }
-
-    @Test
-    void decimalPointOnLeftEndFails() {
-        String value = ".1234";
-        displayHelper.displayValue(value);
-        verify(log).error("Display value must have decimal points appearing strictly between the digits");
-        verify(log, never()).info("Displaying value: {}", ".1234");
-    }
-
-    @Test
-    void decimalPointOnRightEndFails() {
-        String value = "1234.";
-        displayHelper.displayValue(value);
-        verify(log).error("Display value must have decimal points appearing strictly between the digits");
-        verify(log, never()).info("Displaying value: {}", "1234.");
-    }
-
-    @Test
-    void decimalPointsOnBothEndsFails() {
-        String value = ".1234.";
-        displayHelper.displayValue(value);
-        verify(log).error("Display value must have decimal points appearing strictly between the digits");
-        verify(log, never()).info("Displaying value: {}", ".1234.");
-    }
-
-    @Test
-    void invalidCharacterFails() {
-        String value = "G";
-        displayHelper.displayValue(value);
[…]
