+++ pi4micronaut-utils/src/main/java/com/opensourcewithslu/outputdevices/FourDigitSevenSegmentDisplayHelper.java
@@ -1,254 +1,0 @@
-package com.opensourcewithslu.outputdevices;
-
-import com.pi4j.io.gpio.digital.DigitalOutput;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.Map;
-
-public class FourDigitSevenSegmentDisplayHelper {
-    private static Logger log = LoggerFactory.getLogger(FourDigitSevenSegmentDisplayHelper.class);
-
-    private final DigitalOutput sdi;
-    private final DigitalOutput rclk;
-    private final DigitalOutput srclk;
-    private final DigitalOutput digit0;
-    private final DigitalOutput digit1;
-    private final DigitalOutput digit2;
-    private final DigitalOutput digit3;
-
-    private String displayValue;
-    private boolean enabled = false;
-    private final int decimalPoint = 0x7f;
-
-    /**
-     * Mapping of characters to their respective byte representation.
-     * Each byte is a bitset where each bit specifies if a specific segment should be disabled (1) or enabled (0).
-     * Note carefully that the bits are inverted, so 0 means enabled and 1 means disabled.
-     */
-    protected static final Map<Character, Integer> CHAR_BITSETS = Map.ofEntries(
-            Map.entry(' ', 0xff),
-            Map.entry('-', 0x02),
-            Map.entry('0', 0xc0),
-            Map.entry('1', 0xf9),
-            Map.entry('2', 0xa4),
-            Map.entry('3', 0xb0),
-            Map.entry('4', 0x99),
-            Map.entry('5', 0x92),
-            Map.entry('6', 0x82),
-            Map.entry('7', 0xf8),
-            Map.entry('8', 0x80),
-            Map.entry('9', 0x90),
-            Map.entry('A', 0x88),
-            Map.entry('B', 0x83),
-            Map.entry('C', 0xc6),
-            Map.entry('D', 0xA1),
-            Map.entry('E', 0x86),
-            Map.entry('F', 0x8e)
-    );
-
-    /**
-     * Constructor for the FourDigitSevenSegmentDisplayHelper class.
-     *
-     * @param sdi    Serial data input
-     * @param rclk   Register clock
-     * @param srclk  Shift register clock
-     * @param digit0 The first digit
-     * @param digit1 The second digit
-     * @param digit2 The third digit
-     * @param digit3 The fourth digit
-     */
-    //tag::const[]
-    public FourDigitSevenSegmentDisplayHelper(DigitalOutput sdi, DigitalOutput rclk, DigitalOutput srclk, DigitalOutput digit0, DigitalOutput digit1, DigitalOutput digit2, DigitalOutput digit3)
-    //end::const[]
-    {
-        this.sdi = sdi;
-        this.rclk = rclk;
-        this.srclk = srclk;
-        this.digit0 = digit0;
-        this.digit1 = digit1;
-        this.digit2 = digit2;
-        this.digit3 = digit3;
-
-        this.displayValue = "";
-    }
-
-    private void shiftOut(Integer data, boolean decimalPointEnabled) {
-        int value;
-        for (int i = 0; i < 8; i++) { // Loop through each bit in the byte, one for each of the 7 segment and the decimal point
-            if (decimalPointEnabled) {
-                value = (data & (1 << (7 - i)) & this.decimalPoint);
-            } else {
-                value = (data & (1 << (7 - i)) | ~this.decimalPoint);
-            }
-            if (value != 0) {
-                sdi.high(); // Disables segment
-            } else {
-                sdi.low(); // Enables segment
-            }
-            srclk.high();
-            srclk.low();
-        }
-        rclk.high();
-        rclk.low();
-    }
-
-    /**
-     * Clears the display.
-     */
-    //tag::method[]
[…]

[…]
