### 2024-11-18 cc0dc6a
method names, display value parsing, new methods, edit tests to reflect parsing
