+++ pi4micronaut-utils/src/main/java/com/opensourcewithslu/outputdevices/FourDigitSevenSegmentDisplayHelper.java
@@ -21,12 +21,15 @@ public class FourDigitSevenSegmentDisplayHelper {
     private boolean enabled = false;
     private final int decimalPoint = 0x7f;
 
+    private char[] digitValues = new char[4];
+    private boolean[] decimalPoints = new boolean[4];
+
     /**
      * Mapping of characters to their respective byte representation.
      * Each byte is a bitset where each bit specifies if a specific segment should be disabled (1) or enabled (0).
      * Note carefully that the bits are inverted, so 0 means enabled and 1 means disabled.
      */
-    protected static final Map<Character, Integer> CHAR_BITSETS = Map.ofEntries(
+    private static final Map<Character, Integer> CHAR_BITSETS = Map.ofEntries(
             Map.entry(' ', 0xff),
             Map.entry('-', 0x02),
             Map.entry('0', 0xc0),
@@ -70,9 +73,36 @@ public class FourDigitSevenSegmentDisplayHelper {
         this.digit2 = digit2;
         this.digit3 = digit3;
 
-        this.displayValue = "";
+        this.displayValue = "    ";
+        this.parseDisplayValue();
     }
 
+    /**
+     * Helper method to start a second thread continuously updating the display.
+     */
+    private void startDisplayThread() {
+        Thread displayThread = new Thread(() -> {
+            while (enabled) {
+                for (int i = 0; i < 4; i++) {
+                    shiftValueToDigit(i, this.digitValues[i], this.decimalPoints[i]);
+                    try {
+                        Thread.sleep(5); // Adjust the delay as needed
+                    } catch (InterruptedException e) {
+                        Thread.currentThread().interrupt();
+                        log.error("Display thread interrupted", e);
+                    }
+                }
+            }
+        });
+        displayThread.start();
+    }
+
+    /**
+     * Helper method to shift out a byte to the shift register.
+     *
+     * @param data
+     * @param decimalPointEnabled
+     */
     private void shiftOut(Integer data, boolean decimalPointEnabled) {
         int value;
         for (int i = 0; i < 8; i++) { // Loop through each bit in the byte, one for each of the 7 segment and the decimal point
@@ -94,16 +124,13 @@ public class FourDigitSevenSegmentDisplayHelper {
     }
 
     /**
-     * Clears the display.
+     * Helper method to shift out a value to a specific digit.
+     *
+     * @param digit
+     * @param c
+     * @param decimalPoint
      */
-    //tag::method[]
-    public void clear()
-    //end::method[]
-    {
-        displayValue = "";
-    }
-
-    public void setDigit(int digit, char c, boolean decimalPoint) {
+    private void shiftValueToDigit(int digit, char c, boolean decimalPoint) {
         digit0.low();
         digit1.low();
         digit2.low();
@@ -138,6 +165,79 @@ public class FourDigitSevenSegmentDisplayHelper {
     }
 
     /**
+     * Helper method to parse the display value into digit values and decimal points.
+     */
+    private void parseDisplayValue() {
+        this.digitValues = new char[]{' ', ' ', ' ', ' '};
+        this.decimalPoints = new boolean[]{false, false, false, false};
+
+        // Add leading space if the first character is a decimal point
+        if (!this.displayValue.isEmpty() && this.displayValue.charAt(0) == '.') {
+            this.displayValue = " " + this.displayValue;
+        }
+
+        // Replace consecutive decimal points with a space in between
+        while (true) {
+            this.displayValue = this.displayValue.replaceAll("\\.\\.", ". .");
+            // If parsing again would result in no change, we're done and can break
+            if (this.displayValue.replaceAll("\\.\\.", ". .").equals(this.displayValue)) {
+                break;
+            }
+        }
+
+        int idx = 0, pos = 0;
+        while (idx < this.displayValue.length() && pos < 4) {
[…]
