+++ pi4micronaut-utils/src/test/java/com/opensourcewithslu/outputdevices/FourDigitSevenSegmentDisplayHelperTest.java
@@ -1,20 +1,13 @@
 package com.opensourcewithslu.outputdevices;
 
-import com.pi4j.context.ContextProperties;
 import com.pi4j.io.gpio.digital.DigitalOutput;
-import com.pi4j.io.i2c.I2CConfig;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import org.slf4j.Logger;
 
 import static org.mockito.Mockito.*;
-import static org.junit.jupiter.api.Assertions.*;
-
-import com.pi4j.context.Context;
-import com.pi4j.crowpi.components.SevenSegmentComponent;
-
-import java.lang.reflect.Field;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class FourDigitSevenSegmentDisplayHelperTest {
     DigitalOutput sdi = mock(DigitalOutput.class);
@@ -27,68 +20,52 @@ public class FourDigitSevenSegmentDisplayHelperTest {
     FourDigitSevenSegmentDisplayHelper displayHelper = new FourDigitSevenSegmentDisplayHelper(sdi, rclk, srclk, digit0, digit1, digit2, digit3);
     Logger log = mock(Logger.class);
 
-    /*@BeforeEach
-    void setUp() throws Exception {
-        // Mock the Context to return a non-null ContextProperties
-        when(pi4jContext.properties()).thenReturn(contextProperties);
-
-        displayHelper = new FourDigitSevenSegmentDisplayHelper(i2cConfig, pi4jContext);
-
-        // Use reflection to set the mock SevenSegmentComponent
-        Field displayField = FourDigitSevenSegmentDisplayHelper.class.getDeclaredField("display");
-        displayField.setAccessible(true);
-        displayField.set(displayHelper, displayComponent);
-    }*/
-
     @BeforeEach
     public void openMocks() {
         displayHelper.setLog(log);
     }
 
     @Test
-    void longNumberFails() {
+    void longNumberIsCutOff() {
         String value = "12345";
-        displayHelper.displayValue(value);
-        verify(log).error("Display value must not have more than 4 non-decimal point characters");
-        verify(log, never()).info("Displaying value: {}", "12345");
+        displayHelper.print(value);
+        verify(log).info("Displaying value: {}", "1234");
     }
 
     @Test
-    void consecutiveDecimalPointsFails() {
+    void consecutiveDecimalPointIsParsed() {
         String value = "1..23";
-        displayHelper.displayValue(value);
-        verify(log).error("Display value cannot have consecutive decimal points");
-        verify(log, never()).info("Displaying value: {}", "1..23");
+        displayHelper.print(value);
+        String myValue = " ";
+        assertEquals(' ', myValue.charAt(0));
+        verify(log).info("Displaying value: {}", "1. .23");
+
+        String displayed = displayHelper.getDisplayValue();
+        assertEquals("1. .23", displayed);
     }
 
     @Test
-    void decimalPointOnLeftEndFails() {
+    void multipleConsecutiveDecimalPointIsParsed() {
+        String value = "1...3";
+        displayHelper.print(value);
+        verify(log).info("Displaying value: {}", "1. . .3");
+        String displayed = displayHelper.getDisplayValue();
+        assertEquals("1. . .3", displayed);
+    }
+
+    @Test
+    void decimalPointOnLeftEndCutsOff() {
         String value = ".1234";
-        displayHelper.displayValue(value);
-        verify(log).error("Display value must have decimal points appearing strictly between the digits");
-        verify(log, never()).info("Displaying value: {}", ".1234");
-    }
-
-    @Test
-    void decimalPointOnRightEndFails() {
-        String value = "1234.";
-        displayHelper.displayValue(value);
-        verify(log).error("Display value must have decimal points appearing strictly between the digits");
[…]
