### 2024-11-18 de4a69d
New tests, ensuring that display is enabled, finished documentation
