+++ pi4micronaut-utils/src/test/java/com/opensourcewithslu/outputdevices/FourDigitSevenSegmentDisplayHelperTest.java
@@ -26,7 +26,29 @@ public class FourDigitSevenSegmentDisplayHelperTest {
     }
 
     @Test
+    void enable() {
+        displayHelper.enable();
+        verify(log).info("Display enabled");
+    }
+
+    @Test
+    void disable() {
+        displayHelper.enable();
+        verify(log).info("Display enabled");
+        displayHelper.disable();
+        verify(log).info("Display disabled");
+    }
+
+    @Test
+    void printFailsWhenDisabled() {
+        displayHelper.print("1234");
+        verify(log).error("Display must be enabled first");
+        verify(log, never()).info("Displaying value: {}", "1234");
+    }
+
+    @Test
     void longNumberIsCutOff() {
+        displayHelper.enable();
         String value = "12345";
         displayHelper.print(value);
         verify(log).info("Displaying value: {}", "1234");
@@ -34,6 +56,7 @@ public class FourDigitSevenSegmentDisplayHelperTest {
 
     @Test
     void consecutiveDecimalPointIsParsed() {
+        displayHelper.enable();
         String value = "1..23";
         displayHelper.print(value);
         String myValue = " ";
@@ -46,6 +69,7 @@ public class FourDigitSevenSegmentDisplayHelperTest {
 
     @Test
     void multipleConsecutiveDecimalPointIsParsed() {
+        displayHelper.enable();
         String value = "1...3";
         displayHelper.print(value);
         verify(log).info("Displaying value: {}", "1. . .3");
@@ -55,6 +79,7 @@ public class FourDigitSevenSegmentDisplayHelperTest {
 
     @Test
     void decimalPointOnLeftEndCutsOff() {
+        displayHelper.enable();
         String value = ".1234";
         displayHelper.print(value);
         verify(log).info("Displaying value: {}", " .123");
@@ -64,6 +89,7 @@ public class FourDigitSevenSegmentDisplayHelperTest {
 
     @Test
     void invalidCharacterFails() {
+        displayHelper.enable();
         String value = "G";
         displayHelper.print(value);
         verify(log).error("Each display value digit must be numeric, a letter A to F (case insensitive), a hyphen, or a space");
@@ -72,6 +98,7 @@ public class FourDigitSevenSegmentDisplayHelperTest {
 
     @Test
     void displaysLetters() {
+        displayHelper.enable();
         String value = "ABCD";
         displayHelper.print(value);
         verify(log).info("Displaying value: {}", "ABCD");
@@ -82,6 +109,7 @@ public class FourDigitSevenSegmentDisplayHelperTest {
 
     @Test
     void displaysLettersWithLowercase() {
+        displayHelper.enable();
         String value = "abcd";
         displayHelper.print(value);
         verify(log).info("Displaying value: {}", "ABCD");
@@ -92,6 +120,7 @@ public class FourDigitSevenSegmentDisplayHelperTest {
 
     @Test
     void displaysLettersWithMixedCase() {
+        displayHelper.enable();
         String value = "aBcD";
         displayHelper.print(value);
         verify(log).info("Displaying value: {}", "ABCD");
@@ -102,6 +131,7 @@ public class FourDigitSevenSegmentDisplayHelperTest {
 
     @Test
     void displaysHyphen() {
+        displayHelper.enable();
         String value = "-";
         displayHelper.print(value);
         verify(log).info("Displaying value: {}", "-   ");
@@ -112,6 +142,7 @@ public class FourDigitSevenSegmentDisplayHelperTest {
 
     @Test
     void displaysSpaces() {
+        displayHelper.enable();
         String value = "2 2";
         displayHelper.print(value);
         verify(log).info("Displaying value: {}", "2 2 ");
[…]
