+++ pi4micronaut-utils/src/docs/asciidoc/components/outputComponents/fourDigitSevenSegment.adoc
@@ -123,11 +123,49 @@ $ curl http://localhost:8080/four-digit-seven-segment/displayValue/1%2034
 
 [source,yaml]
 ----
-i2c:
-four-digit-seven-segment:
-  name: 4 Digit 7 Segment Display
-  bus: 1
-  device: 0x27
+digital-output:
+  digit-0:
+    name: Digit 0
+    address: 17
+    shutdown: LOW
+    initial: LOW
+    provider: pigpio-digital-output
+  digit-1:
+    name: Digit 1
+    address: 27
+    shutdown: LOW
+    initial: LOW
+    provider: pigpio-digital-output
+  digit-2:
+    name: Digit 2
+    address: 22
+    shutdown: LOW
+    initial: LOW
+    provider: pigpio-digital-output
+  digit-3:
+    name: Digit 3
+    address: 10
+    shutdown: LOW
+    initial: LOW
+    provider: pigpio-digital-output
+  sdi:
+    name: SDI
+    address: 24
+    shutdown: LOW
+    initial: LOW
+    provider: pigpio-digital-output
+  rclk:
+    name: RCLK
+    address: 23
+    shutdown: LOW
+    initial: LOW
+    provider: pigpio-digital-output
+  srclk:
+    name: SRCLK
+    address: 18
+    shutdown: LOW
+    initial: LOW
+    provider: pigpio-digital-output
 ----
 
 ===== Constructor and Methods
