### 2024-11-18 c85e0d8
Helper changes
