### 2024-11-24 71d20a0
Added servo motor helper tests
