### 2024-11-24 615c787
modified servo motor disable() test
