+++ pi4micronaut-utils/src/test/java/com/opensourcewithslu/outputdevices/ServoMotorHelperTest.java
@@ -30,6 +30,7 @@ public class ServoMotorHelperTest {
 
     @Test
     void disables() {
+        servoMotorHelper.enable();
         servoMotorHelper.disable();
         verify(servoMotor).off();
         verify(log).info("Disabling servo motor");
