### 2024-11-25 aa0ae2b
Documentation for Thermistor Sensor
