+++ pi4micronaut-utils/src/docs/asciidoc/components/inputComponents/thermistor.adoc
@@ -1,0 +1,80 @@
+:imagesdir: img/
+
+ifndef::rootpath[]
+:rootpath: ../../
+endif::rootpath[]
+
+ifdef::rootpath[]
+:imagesdir: {rootpath}{imagesdir}
+endif::rootpath[]
+
+==== Thermistor Sensor
+[.text-right]
+https://github.com/oss-slu/Pi4Micronaut/edit/develop/pi4micronaut-utils/src/docs/asciidoc/components/inputComponents/thermistor.adoc
+
+
+===== Overview
+
+This section provides details of the Thermistor, including the components and assembly instructions.
+
+===== Components
+
+* Thermistor
+* Breadboard
+* T-Extension Board
+* ADC0834 Analog to Digital Converter
+* Jumper wires x 16
+* Power source (5V)
+* 10KΩ resistor
+
+===== Assembly Instructions
+
+. 
+
+===== Wiring Diagram
+
+image::thermistor_schematic.webp[]
+
+
+===== Circuit Diagram
+
+
+
+===== Functionality
+
+A Thermistor -  A thermistor is a thermally sensitive resistor that exhibits a precise and predictable change in resistance proportional to small changes in temperature. 
+
+===== Testing
+Use the below command to get Fahrenheit.
+[source, bash]
+----
+$ curl http://localhost:8080/temperature/fahrenheit
+----
+
+Use the below command to get Celsius.
+[source, bash]
+----
+$ curl http://localhost:8080/temperature/celsius
+----
+
+===== Troubleshooting
+- Sensor not detecting something:
+    * 
+
+
+===== YAML Configuration
+[source, yaml]
+
+----
+digital-input:
+    
+----
+
+===== Constructor and Methods
+The constructor and the methods within the ThermistorHelper class can be seen in our javadoc link:https://oss-slu.github.io/Pi4Micronaut/javadoc/com/opensourcewithslu/inputdevices/ThermistorHelper.html[here].
+
+===== Example Controller
+
+====== This controller uses the thermistor to detect a temperature value
+[source, java]
+include::../../../../../../components/src/main/java/com/opensourcewithslu/components/controllers/ThermistorController.java[tag=ex]
