### 2024-11-25 36bf88c
Updated Adoc to have only Wiring Diagram
