### 2024-11-25 0ff5c06
added the requirments for issue 293 Sprint5
