+++ pi4micronaut-utils/src/docs/asciidoc/components/outputComponents/SevenSegementDisplay.adoc
@@ -1,0 +1,136 @@
+:imagesdir: img/
+
+ifndef::rootpath[]
+:rootpath: ../../
+endif::rootpath[]
+
+ifdef::rootpath[]
+:imagesdir: {rootpath}{imagesdir}
+endif::rootpath[]
+
+==== Seven Segment Display
+
+[.text-right]
+https://github.com/oss-slu/Pi4Micronaut/edit/develop/pi4micronaut-utils/src/docs/asciidoc/components/outputComponents/SevenSegmentDisplay.adoc[Improve this doc]
+
+===== Overview
+This section provides comprehensive details about the Seven Segment Display component, covering its configuration, circuit setup, and usage instructions.
+
+===== Components
+* 1 x Raspberry Pi
+* 1 x Seven Segment Display
+* Jumper wires
+* Breadboard
+* Power source
+
+===== Circuit Setup
+
+1. Place the Seven Segment Display on the breadboard.
+2. Connect each segment of the display to its respective GPIO pin as specified in the YAML configuration.
+3. Use a common ground to connect the display to the Raspberry Pi.
+4. Refer to the circuit diagram for detailed wiring instructions.
+
+===== Circuit Diagram
+
+image::/Users/jeongseyun7/Desktop/Capston1/Sprint4/Refreshed/Pi4Micronaut/Circuit_Diagram.png[]
+
+===== Schematic Diagram
+
+image::/Users/jeongseyun7/Desktop/Capston1/Sprint4/Refreshed/Pi4Micronaut/Schematic Diagram.png[]
+
+===== YAML Configuration
+The following YAML configuration is used to set up the Seven Segment Display in `application.yml`. This configuration defines each segment as a digital output and includes the I2C parameters necessary for Raspberry Pi integration.
+
+[source, yaml]
+----
+  # tag::i2c[]
+  i2c:
+    seven-segment-display:
+      name: "Single Seven-Segment Display"
+      bus: 1
+      device: 0x3F
+      brightness: 100
+      decimal-point: true
+      segments:
+        digital-output:
+          segment-a:
+            name: Segment A
+            address: 2
+            shutdown: LOW
+            initial: LOW
+            provider: pigpio-digital-output
+          segment-b:
+            name: Segment B
+            address: 3
+            shutdown: LOW
+            initial: LOW
+            provider: pigpio-digital-output
+          segment-c:
+            name: Segment C
+            address: 4
+            shutdown: LOW
+            initial: LOW
+            provider: pigpio-digital-output
+          segment-d:
+            name: Segment D
+            address: 5
+            shutdown: LOW
+            initial: LOW
+            provider: pigpio-digital-output
+          segment-e:
+            name: Segment E
+            address: 6
+            shutdown: LOW
+            initial: LOW
+            provider: pigpio-digital-output
+          segment-f:
+            name: Segment F
+            address: 7
+            shutdown: LOW
+            initial: LOW
+            provider: pigpio-digital-output
+          segment-g:
+            name: Segment G
+            address: 8
+            shutdown: LOW
+            initial: LOW
+            provider: pigpio-digital-output
+          segment-dot:
+            name: Segment DOT
+            address: 9
+            shutdown: LOW
+            initial: LOW
+            provider: pigpio-digital-output
+  # end::i2c[]
+----
+
+===== Functionality
+The Seven Segment Display can display numbers from 0-9 by controlling each segment. Additionally, the decimal point can be activated as needed.
+
+===== Testing
+
+To test the Seven Segment Display component with the Micronaut API, use the following commands:
+
+[source, bash]
+----
+$ curl http://localhost:8080/sevensegment/display/5
+----
+* `/display/{number}` - Displays the specified number (0-9).
+* `/reset` - Turns off all segments on the display.
[…]
