+++ pi4micronaut-utils/src/main/resources/application.yml
@@ -5,3 +5,62 @@ netty:
   default:
     allocator:
       max-order: 3
+i2c:
+  seven-segment-display:
+    name: "Single Seven-Segment Display"
+    bus: 1
+    device: 0x3F
+    brightness: 100
+    decimal-point: true
+    segments:
+      digital-output:
+        # Define each segment of the seven-segment display as a digital output.
+        segment-a:
+          name: Segment A
+          address: 2
+          shutdown: LOW
+          initial: LOW
+          provider: pigpio-digital-output
+        segment-b:
+          name: Segment B
+          address: 3
+          shutdown: LOW
+          initial: LOW
+          provider: pigpio-digital-output
+        segment-c:
+          name: Segment C
+          address: 4
+          shutdown: LOW
+          initial: LOW
+          provider: pigpio-digital-output
+        segment-d:
+          name: Segment D
+          address: 5
+          shutdown: LOW
+          initial: LOW
+          provider: pigpio-digital-output
+        segment-e:
+          name: Segment E
+          address: 6
+          shutdown: LOW
+          initial: LOW
+          provider: pigpio-digital-output
+        segment-f:
+          name: Segment F
+          address: 7
+          shutdown: LOW
+          initial: LOW
+          provider: pigpio-digital-output
+        segment-g:
+          name: Segment G
+          address: 8
+          shutdown: LOW
+          initial: LOW
+          provider: pigpio-digital-output
+        segment-dot:
+          name: Segment DOT
+          address: 9
+          shutdown: LOW
+          initial: LOW
+          provider: pigpio-digital-output
+      # end::i2c[]
