### 2024-11-25 8c05dee
Added the documentation for Fancontroller
