+++ pi4micronaut-utils/src/docs/asciidoc/components/outputComponents/FanController.adoc
@@ -1,0 +1,90 @@
+:imagesdir: img/
+
+ifndef::rootpath[]
+:rootpath: ../../
+endif::rootpath[]
+
+ifdef::rootpath[]
+:imagesdir: {rootpath}{imagesdir}
+endif::rootpath[]
+
+==== Fan Controller
+
+[.text-right]
+https://github.com/oss-slu/Pi4Micronaut/edit/develop/pi4micronaut-utils/src/docs/asciidoc/components/outputComponents/FanController.adoc[Improve this doc]
+
+===== Overview
+This section provides comprehensive details about the Fan Controller component, covering its configuration, circuit setup, and usage instructions.
+
+===== Components
+* 1 x Raspberry Pi
+* 1 x Fan (PWM-controlled or simple DC fan)
+* Jumper wires
+* Breadboard
+* Power source
+
+===== Circuit Setup
+
+1. Connect the fan to the Raspberry Pi GPIO pins as specified in the YAML configuration.
+2. Use a PWM-compatible pin to control the fan speed (if the fan supports PWM).
+3. Connect a common ground between the fan and the Raspberry Pi.
+4. Refer to the circuit diagram for detailed wiring instructions.
+
+===== Circuit Diagram
+
+image::/Users/jeongseyun7/Desktop/Capston1/Sprint4/Refreshed/Pi4Micronaut/Circuit_Diagram1.png[]
+
+===== Schematic Diagram
+
+image::/Users/jeongseyun7/Desktop/Capston1/Sprint4/Refreshed/Pi4Micronaut/Schematic_Diagram1.png[]
+
+===== YAML Configuration
+The following YAML configuration is used to set up the Fan Controller in `application.yml`. This configuration includes parameters to control fan speed and behavior.
+
+[source, yaml]
+----
+  # tag::fan[]
+  fan:
+    controller:
+      name: "Raspberry Pi Fan Controller"
+      pwm-pin:
+        name: Fan PWM Pin
+        address: 18
+        frequency: 25000
+        shutdown: LOW
+        initial: LOW
+        provider: pigpio-pwm-output
+      speed-settings:
+        low: 30
+        medium: 60
+        high: 100
+  # end::fan[]
+----
+
+===== Functionality
+The Fan Controller allows the Raspberry Pi to control the speed of a fan using PWM (Pulse Width Modulation) or simple on/off states. It supports predefined speed levels (e.g., low, medium, high).
+
+===== Testing
+
+To test the Fan Controller component with the Micronaut API, use the following commands:
+
+[source, bash]
+----
+$ curl http://localhost:8080/fan/speed/high
+----
+* `/speed/{level}` - Sets the fan speed to the specified level (`low`, `medium`, or `high`).
+* `/stop` - Turns off the fan.
+* `/shutdown` - Deactivates the fan completely and resets settings.
+
+===== Troubleshooting
+- **Fan not working**: Verify the fan is connected to the correct GPIO pins and ensure power is supplied to the fan.
+- **Inconsistent speed**: Review the PWM frequency in the YAML configuration and ensure compatibility with the fan.
+
+===== Example Controller
+
+The following example controller demonstrates how to set up routes for controlling fan speed, stopping the fan, and managing shutdown features.
+
+[source, java]
+----
+include::../../../../../../components/src/main/java/com/opensourcewithslu/components/controllers/FanController.java[tag=ex]
+----
