### 2024-11-29 29d1c70
Added try-catch
