### 2024-11-29 708ce8c
Edited try-catch
