+++ pi4micronaut-utils/src/main/java/com/opensourcewithslu/outputdevices/SevenSegmentDisplayHelper.java
@@ -1,16 +1,14 @@
 package com.opensourcewithslu.outputdevices;
 
-import com.pi4j.context.Context;
 import com.pi4j.io.gpio.digital.DigitalOutput;
-import com.pi4j.io.gpio.digital.DigitalState;
-import com.pi4j.Pi4J;
-import io.micronaut.context.annotation.Value;
-import jakarta.inject.Singleton;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-@Singleton
+import java.util.Map;
+
 public class SevenSegmentDisplayHelper {
+    private static Logger log = LoggerFactory.getLogger(SevenSegmentDisplayHelper.class);
 
-    private final Context pi4j;
     private final DigitalOutput pinA;
     private final DigitalOutput pinB;
     private final DigitalOutput pinC;
@@ -20,102 +18,220 @@ public class SevenSegmentDisplayHelper {
     private final DigitalOutput pinG;
     private final DigitalOutput decimalPoint;
 
-    @Value("${i2c.seven-segment-display.segments.digital-output.segment-a.address}")
-    int pinAAddress;
-    @Value("${i2c.seven-segment-display.segments.digital-output.segment-b.address}")
-    int pinBAddress;
-    @Value("${i2c.seven-segment-display.segments.digital-output.segment-c.address}")
-    int pinCAddress;
-    @Value("${i2c.seven-segment-display.segments.digital-output.segment-d.address}")
-    int pinDAddress;
-    @Value("${i2c.seven-segment-display.segments.digital-output.segment-e.address}")
-    int pinEAddress;
-    @Value("${i2c.seven-segment-display.segments.digital-output.segment-f.address}")
-    int pinFAddress;
-    @Value("${i2c.seven-segment-display.segments.digital-output.segment-g.address}")
-    int pinGAddress;
-    @Value("${i2c.seven-segment-display.segments.digital-output.segment-dot.address}")
-    int decimalPointPinAddress;
+    private char digitValue;
+    private boolean decimalPointEnabled = false;
+    private boolean enabled = false;
 
-    // Segment pattern configuration for displaying numbers 0-9
-    private static final boolean[][] DIGIT_SEGMENTS = {
-            {true, true, true, true, true, true, false}, // 0
-            {false, true, true, false, false, false, false}, // 1
-            {true, true, false, true, true, false, true}, // 2
-            {true, true, true, true, false, false, true}, // 3
-            {false, true, true, false, false, true, true}, // 4
-            {true, false, true, true, false, true, true}, // 5
-            {true, false, true, true, true, true, true}, // 6
-            {true, true, true, false, false, false, false}, // 7
-            {true, true, true, true, true, true, true}, // 8
-            {true, true, true, true, false, true, true}  // 9
-    };
+    /**
+     * Mapping of characters to their respective byte representation.
+     * Each byte is a bitset where each bit specifies if a specific segment should be disabled (1) or enabled (0).
+     * Note carefully that the bits are inverted, so 0 means enabled and 1 means disabled.
+     */
+    private static final Map<Character, Integer> CHAR_BITSETS = Map.ofEntries(
+            Map.entry(' ', 0xff),
+            Map.entry('-', 0x02),
+            Map.entry('0', 0xc0),
+            Map.entry('1', 0xf9),
+            Map.entry('2', 0xa4),
+            Map.entry('3', 0xb0),
+            Map.entry('4', 0x99),
+            Map.entry('5', 0x92),
+            Map.entry('6', 0x82),
+            Map.entry('7', 0xf8),
+            Map.entry('8', 0x80),
+            Map.entry('9', 0x90),
[…]
