+++ pi4micronaut-utils/src/docs/asciidoc/components/inputComponents/thermistor.adoc
@@ -9,16 +9,15 @@ ifdef::rootpath[]
 endif::rootpath[]
 
 ==== Thermistor Sensor
+
 [.text-right]
 https://github.com/oss-slu/Pi4Micronaut/edit/develop/pi4micronaut-utils/src/docs/asciidoc/components/inputComponents/thermistor.adoc
-
 
 ===== Overview
 
 This section provides details of the Thermistor, including the components and assembly instructions.
 
 ===== Components
-
 * Thermistor
 * Breadboard
 * T-Extension Board
@@ -29,9 +28,15 @@ This section provides details of the Thermistor, including the components and as
 
 ===== Assembly Instructions
 
-. 
+. Insert the ADC0834 chip into the breadboard with its pins centered around the central divider
+. Connect VCC (pin 16) of ADC0834 to the 5V pin on the Raspberry Pi GPIO.
+. Connect GND (pin 8) to the GND pin on the Raspberry Pi GPIO.
+. CLK (pin 18): Connect to the (GPIO18) pin on the Raspberry Pi.
+. DOUT (pin 27): Connect to the (GPIO27) pin on the Raspberry Pi.
+. DIN (pin 27): Connect to the (GPIO10) pin on the Raspberry Pi.
+. CS (pin 17): Connect to the (GPIO17) pin on the Raspberry Pi.
 
-===== Wiring Diagram
+===== Wiring Diagrams
 
 image::thermistor_schematic.webp[]
 
@@ -54,6 +59,7 @@ $ curl http://localhost:8080/temperature/celsius
 
 ===== Troubleshooting
 - Sensor not detecting something:
+    * Make sure the pins of the thermistor are in the correct pins of ADC0834
     * 
 
 
@@ -62,6 +68,7 @@ $ curl http://localhost:8080/temperature/celsius
 
 ----
 digital-input:
+
     
 ----
 
