+++ components/src/main/resources/application.yml
@@ -242,50 +242,50 @@ pi4j:
       initial: LOW
       provider: pigpio-digital-output
     pinA:
-      name: PINA
-      address: 30
+      name: pinA
+      address: 23
       shutdown: LOW
       initial: LOW
       provider: pigpio-digital-output
     pinB:
-      name: PINB
-      address: 31
+      name: pinB
+      address: 24
       shutdown: LOW
       initial: LOW
       provider: pigpio-digital-output
     pinC:
-      name: PINC
-      address: 32
+      name: pinC
+      address: 5
       shutdown: LOW
       initial: LOW
       provider: pigpio-digital-output
     pinD:
-      name: PIND
-      address: 33
+      name: pinD
+      address: 26
       shutdown: LOW
       initial: LOW
       provider: pigpio-digital-output
     pinE:
-      name: PINE
-      address: 34
+      name: pinE
+      address: 6
       shutdown: LOW
       initial: LOW
       provider: pigpio-digital-output
     pinF:
-      name: PINF
-      address: 35
+      name: pinF
+      address: 16
       shutdown: LOW
       initial: LOW
       provider: pigpio-digital-output
     pinG:
-      name: PING
-      address: 36
+      name: pinG
+      address: 12
       shutdown: LOW
       initial: LOW
       provider: pigpio-digital-output
     decimalPoint:
-      name: Decimalpoint
-      address: 37
+      name: decimalPoint
+      address: 13
       shutdown: LOW
       initial: LOW
       provider: pigpio-digital-output
