+++ pi4micronaut-utils/src/main/java/com/opensourcewithslu/outputdevices/SevenSegmentDisplayHelper.java
@@ -1,237 +1,0 @@
-package com.opensourcewithslu.outputdevices;
-
-import com.pi4j.io.gpio.digital.DigitalOutput;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.Map;
-
-public class SevenSegmentDisplayHelper {
-    private static Logger log = LoggerFactory.getLogger(SevenSegmentDisplayHelper.class);
-
-    private final DigitalOutput pinA;
-    private final DigitalOutput pinB;
-    private final DigitalOutput pinC;
-    private final DigitalOutput pinD;
-    private final DigitalOutput pinE;
-    private final DigitalOutput pinF;
-    private final DigitalOutput pinG;
-    private final DigitalOutput decimalPoint;
-
-    private char digitValue;
-    private boolean decimalPointEnabled = false;
-    private boolean enabled = false;
-
-    /**
-     * Mapping of characters to their respective byte representation.
-     * Each byte is a bitset where each bit specifies if a specific segment should be disabled (1) or enabled (0).
-     * Note carefully that the bits are inverted, so 0 means enabled and 1 means disabled.
-     */
-    private static final Map<Character, Integer> CHAR_BITSETS = Map.ofEntries(
-            Map.entry(' ', 0xff),
-            Map.entry('-', 0x02),
-            Map.entry('0', 0xc0),
-            Map.entry('1', 0xf9),
-            Map.entry('2', 0xa4),
-            Map.entry('3', 0xb0),
-            Map.entry('4', 0x99),
-            Map.entry('5', 0x92),
-            Map.entry('6', 0x82),
-            Map.entry('7', 0xf8),
-            Map.entry('8', 0x80),
-            Map.entry('9', 0x90),
-            Map.entry('A', 0x88),
-            Map.entry('B', 0x83),
-            Map.entry('C', 0xc6),
-            Map.entry('D', 0xa1),
-            Map.entry('E', 0x86),
-            Map.entry('F', 0x8e)
-    );
-
-    /**
-     * Constructor for the SevenSegmentDisplayHelper class.
-     *
-     * @param pinA Segment A
-     * @param pinB Segment B
-     * @param pinC Segment C
-     * @param pinD Segment D
-     * @param pinE Segment E
-     * @param pinF Segment F
-     * @param pinG Segment G
-     * @param decimalPoint The decimal point
-     */
-    //tag::const[]
-    public SevenSegmentDisplayHelper(DigitalOutput pinA, DigitalOutput pinB, DigitalOutput pinC, DigitalOutput pinD, DigitalOutput pinE, DigitalOutput pinF, DigitalOutput pinG, DigitalOutput decimalPoint)
-    //end::const[]
-    {
-        this.pinA = pinA;
-        this.pinB = pinB;
-        this.pinC = pinC;
-        this.pinD = pinD;
-        this.pinE = pinE;
-        this.pinF = pinF;
-        this.pinG = pinG;
-        this.decimalPoint = decimalPoint;
-
-        this.digitValue = ' ';
-    }
-
-    /**
-     * Helper method to start a second thread continuously updating the display.
-     */
-    private void startDisplayThread() {
-        Thread displayThread = new Thread(() -> {
-            while (enabled) {
-                shiftValueToDigit(this.digitValue, this.decimalPointEnabled);
-                try {
-                    Thread.sleep(5); // Adjust the delay as needed
-                } catch (InterruptedException e) {
-                    Thread.currentThread().interrupt();
-                    log.error("Display thread interrupted", e);
-                }
-            }
-        });
-        displayThread.start();
-    }
-
-    /**
-     * Helper method to shift out a byte to the shift register.
-     *
[…]
