### 2024-12-19 #7486 @majensen
Hello @dhhagan, would you be willing to review [this work](#7486) for [JOSS](https://joss.theoj.org)? I would appreciate your expertise. Thanks --@majensen
