### 2025-02-01 37d107d
test(slide-switch): add comprehensive unit tests for SlideSwitchHelper
