+++ pi4micronaut-utils/src/main/java/com/opensourcewithslu/inputdevices/SlideSwitchHelper.java
@@ -9,7 +9,7 @@ import org.slf4j.LoggerFactory;
  * The SlideSwitchHelper class is used to initialize a slide switch.
  */
 public class SlideSwitchHelper {
-    private static final Logger log = LoggerFactory.getLogger(SlideSwitchHelper.class);
+    private static Logger log = LoggerFactory.getLogger(SlideSwitchHelper.class);
 
     private final DigitalInput slideSwitchInput;
 
@@ -41,9 +41,26 @@ public class SlideSwitchHelper {
     {
         log.info("Initializing Slide Switch");
 
-        slideSwitchInput.addListener(e->
-            isOn = slideSwitchInput.isHigh()
-        );
+        // Update the current state and log the appropriate message
+        isOn = slideSwitchInput.isHigh();
+        if (isOn) {
+            log.info("Turning Switch On");
+        } else {
+            log.info("Turning Switch Off");
+        }
+
+        // Add a single listener that updates the state and logs on change
+        slideSwitchInput.addListener(e-> {
+            boolean currentState = slideSwitchInput.isHigh();
+            if (currentState != isOn) {
+                isOn = currentState;
+                if (isOn) {
+                    log.info("Turning Switch On");
+                } else {
+                    log.info("Turning Switch Off");
+                }
+            }
+        });
     }
 
     /**
@@ -67,4 +84,10 @@ public class SlideSwitchHelper {
     {
         slideSwitchInput.removeListener(function);
     }
+
+    /**
+     *
+     * @param log Logger object to set the logger to.
+     */
+    public void setLog(Logger log) { this.log = log; }
 }
