### 2025-02-01 #7486 @majensen
Hi @lucask07 - would you be willing to review [this work](#7486) for [JOSS](https://joss.theoj.org)? Thanks for reviewing for us in the past; I can really use your expertise here. Thanks! --@majensen
