### 2025-02-01 #7486 @majensen
Hello @alustig3 -- would you have time and interest in reviewing [this work](#7486) for [JOSS](https://joss.theoj.org)? I appreciate your reviewing in the past, and I think this one is right up your alley. Thanks for considering! --@majensen
