### 2025-02-01 #7486 @majensen
Hi @tuliofalmeida - would you be able to review [this work](#7486) for [JOSS](https://joss.theoj.org)? We'd really appreciate your expertise. Thanks for your consideration! --@majensen
