### 2025-02-02 d390edc
Added unit tests for MicroSwitchController class
