+++ pi4micronaut-utils/src/main/java/com/opensourcewithslu/inputdevices/MicroSwitchHelper.java
@@ -10,7 +10,7 @@ import org.slf4j.LoggerFactory;
  */
 public class MicroSwitchHelper {
 
-    private static final Logger log = LoggerFactory.getLogger(MicroSwitchHelper.class);
+    private static Logger log = LoggerFactory.getLogger(MicroSwitchHelper.class);
 
     private final DigitalInput microSwitchInput;
 
@@ -64,4 +64,9 @@ public class MicroSwitchHelper {
             microSwitchInputListener = null;
         }
     }
+
+    public void setLog(Logger log) {
+        this.log = log;
+    }
+
 }
