### 2025-02-03 #7486 @majensen
Hello @shahchiragh -- would you have time to take on an additional review (#7486) for [JOSS](https://joss.theoj.org)? The topic is specialized - I'd really appreciate your expertise. Thanks for considering --@majensen
