### 2025-02-03 #7486 @majensen
@shahchiragh fantastic! It would be great to shoot for 3 weeks. Does that work for you?
