### 2025-02-04 #7744 @editorialbot
[REVIEW]: Pi4Micronaut: A Java Library for IoT Research Instruments on Raspberry Pi
**Submitting author:** @ruthvikm (<a href="http://orcid.org/0009-0004-9943-7559">Ruthvik Mannem</a>)
**Repository:** https://github.com/oss-slu/Pi4Micronaut
**Branch with paper.md** (empty if default branch): paper-submission
**Version:** v1.1
