### 2025-02-04 #7744 @abhishektiwari
@ruthvikm - Can you please clarify authorship of this paper? I see 16 contributors to [Pi4Micronaut](https://github.com/oss-slu/Pi4Micronaut) Github project  but only 2 authors on the JOSS paper. Can you confirm if list of authors is appropriate and complete? @majensen Please confirm with EiC if this is acceptable.
