### 2025-02-09 a05ecac
Undo changes from SlideSwitchController and SlideSwitchHelper and only added SlideSwitchHelperTest to test out the unit test
