### 2025-02-15 5f1e730
test: add comprehensive unit tests for PIRSensorHelper
