### 2025-02-16 8b78f10
Photo Resistor Helper Unit Tests Added
