### 2025-02-18 d91c04d
Update PhotoResistorHelperTest.java
made changes to the edge case testing to get rid of unnecessary delays. 

Signed-off-by: avaenk <ava.enke@slu.edu>
