### 2025-02-18 a40809e
Update PhotoResistorHelperTest.java
Signed-off-by: avaenk <ava.enke@slu.edu>
