### 2025-02-18 8190eb3
Update PhotoResistorHelperTest.java
Signed-off-by: avaenk <ava.enke@slu.edu>
