### 2025-02-18 bb0b292
Update PhotoResistorHelperTest.java
Signed-off-by: avaenk <ava.enke@slu.edu>
