### 2025-02-24 6bf6d33
Corrected Documentation and UI Bug
