### 2025-02-25 550a954
Upgrading micronaut version to 4.7.6
