### 2025-02-27 #7744 @ruthvikm
@abhishektiwari - Sorry for the delayed response. This is an open-source project, and we are continuously extending the library to support more electronic components, so we expect more contributors over time. I was responsible for releasing versions v1.0 and v1.1. Me and Adrian were the primary contributors to the JOSS paper, which is why only our names are mentioned as authors.
