### 2025-03-01 832f8dc
Added the configuration classes
