### 2025-03-02 512c5b9
Reviewed PushButtonHelper for Micronaut 4.7.6 compatibility – No issues found, awaiting hardware testing
