### 2025-03-02 97b8c45
Upgrade to Micronaut 4.7.6 and update DigitalOutputConfiguration binding with supporting unit tests.
