### 2025-03-03 72a3e6d
Upgrade to Micronaut 4.7.6 and update DigitalOutputConfiguration binding with supporting unit tests.
