### 2025-03-03 03bc5f3
Reviewed PushButtonHelper for Micronaut 4.7.6 compatibility – No issu… …es found, awaiting hardware testing
