### 2025-03-09 #7744 @abhishektiwari
Thanks @ruthvikm. Please see my initial feedback.

## Software paper
* **Citation syntax for the paper** - Rather than using '@Longley:2017' please use [@Longley:2017]. See citation style guide for [Pandoc here](https://pandoc.org/MANUAL.html#extension-citations).

* **State of the field** - I strongly recommend expanding the the statement of need section to include a few lines on current state of the filed. For example how Pi4Micronaut differs from other high-level Java abstractions for Raspberry Pi’s GPIO pins such as Diozero or Spring for Raspberry Pi.

* **Quality of writing** - small grammatical improvement remove `to be` on Line 18 "Scientific research often requires ~~to be~~ customized data-gathering instruments that can be integrated with a computing platform"
