+++ pi4micronaut-utils/src/main/java/com/opensourcewithslu/utilities/DigitalOutputConfiguration.java
@@ -1,13 +1,14 @@
 package com.opensourcewithslu.utilities;
 
 import com.pi4j.io.gpio.digital.DigitalState;
-import io.micronaut.context.annotation.ConfigurationInject;
 import io.micronaut.context.annotation.EachProperty;
 import io.micronaut.context.annotation.Parameter;
+import io.micronaut.context.annotation.Prototype;
 
 /**
  * This class handles the configuration of a digital output component.
  */
+@Prototype
 @EachProperty("pi4j.digital-output")
 public class DigitalOutputConfiguration {
 
@@ -22,7 +23,6 @@ public class DigitalOutputConfiguration {
      * The DigitalOutputConfiguration constructor.
      * @param id The configuration id as defined in the application.yml.
      */
-    @ConfigurationInject
     public DigitalOutputConfiguration(@Parameter String id) {
         this.id = id;
     }
