### 2025-03-13 #7744 @shahchiragh
Hi @ruthvikm, here is my feedback:

- It would be helpful to add the **"Usage"** section in the project _.readme_ or have a brief mention with "Getting started" steps and then introduce documentation link to [Link to Pi4Micronaut Documentation](https://oss-slu.github.io/Pi4Micronaut/). At first, I was struggling to find the correct docs link.
- Pi4Micronaut is not explicitly compared to other commonly used software packages for Raspberry Pi GPIO control. To strengthen that discussion, I would suggest authors to mention alternative solutionsand highlight how Pi4Micronaut differs or improves upon them. For instance: "Unlike RPi.GPIO, which is a Python-based library, Pi4Micronaut offers a Java-based approach, making it more accessible to researchers who prefer Java for scientific computing.", etc. Additionally, the authors could elaborate on any performance benefits , usability improvements, or additional features (if any) that Pi4Micronaut provides compared to existing tools. This will help with _State of the Field_  section.
- The citation format in the text is incorrect per Pandoc's syntax. Pandoc citations should use square brackets and omit the apostrophe. Correct Format: e.g. [@Knoll2010] "...a well-established programming language in scientific computing [@Knoll2010], to interact with hardware components..."
- I don't see any references, are there not or they are missing?

Let me know if you have any questions.
