+++ pi4micronaut-utils/src/main/java/com/opensourcewithslu/outputdevices/ServoMotorHelper.java
@@ -107,3 +107,4 @@ public class ServoMotorHelper {
         return isEnabled;
     }
 }
+//completed
