### 2025-03-22 #7744 @abhishektiwari
Thanks @ruthvikm. I suggest to add references for Diozero and Spring for Raspberry Pi as mentioned in text.

@majensen complete the review and checklist.
