### 2025-03-30 2e701cf
Fixed issues and added TiltSwitchHelperTests
