### 2025-03-30 ce1a35c
Revert "Add MicroSwitchController and MicroSwitchHelper"
