### 2025-04-05 #7744 @majensen
@ruthvikm I got slightly ahead of myself - I see @shahchiragh still has an open comment regarding comparison of your package to others (https://github.com/openjournals/joss-reviews/issues/7744#issuecomment-2721683160). Can you please address this comment? Feel free to create the archive.
