### 2025-04-07 2221542
upgrade UltraSonicSensor class and added the unit test
