+++ pi4micronaut-utils/src/main/java/com/opensourcewithslu/inputdevices/UltraSonicSensorHelper.java
@@ -1,0 +1,155 @@
+package main.java.com.opensourcewithslu.inputdevices;
+
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
+import java.util.logging.Logger;
+
+import com.pi4j.io.gpio.digital.*;
+
+/**
+ * The UltraSonicSensorHelper class initializes the Ultra Sonic Sensor and provides component functionality
+ */
+public class UltraSonicSensorHelper {
+
+    private static final Logger log = Logger.getLogger(UltraSonicSensorHelper.class.getName());
+    private final DigitalOutput triggerPin; // This is the Trigger Pin - we send the pulse
+    private final DigitalInput echoPin;     // This is the Echo Pin - we read the return signal
+    private volatile boolean sensorActive;
+    private volatile double distance;
+    private final ScheduledExecutorService executorService = Executors.newSingleThreadScheduledExecutor();
+
+    /**
+     * Constructs a new UltraSonicSensorHelper instance.
+     *
+     * @param triggerPin The DigitalOutput pin for triggering the ultrasonic sensor.
+     * @param echoPin The DigitalInput pin for receiving the echo from the ultrasonic sensor.
+     */
+    //tag::const[]
+    public UltraSonicSensorHelper(DigitalOutput triggerPin, DigitalInput echoPin)
+    //end::const[]
+    {
+        this.triggerPin = triggerPin;
+        this.echoPin = echoPin;
+        initialize();
+    }
+
+    /**
+     * Initializes the Ultrasonic Sensor
+     */
+    //tag::method[]
+    public void initialize()
+    //end::method[]
+    {
+        log.info("Ultrasonic Sensor Initialized");
+        sensorActive = true;
+        // TriggerPin is low initially
+        triggerPin.low();
+    }
+
+    /**
+     * Begins measuring distance from sensor calling triggerAndMeasureDistance function every 100 milliseconds
+     */
+    //tag::method[]
+    public void startMeasuring()
+    //end::method[]
+    {
+        if (!sensorActive) {
+            return;
+        }
+        executorService.scheduleAtFixedRate(this::triggerAndMeasureDistance, 0, 100, TimeUnit.MILLISECONDS);
+    }
+
+    /**
+     * Measures distance from ultrasonic sensor to a surface
+     */
+    private void triggerAndMeasureDistance() {
+        try {
+            // Trigger pin is low for a short period before sending the pulse
+            triggerPin.low();
+            TimeUnit.MICROSECONDS.sleep(2);
+
+            // Sending a 10-microsecond pulse on the trigger pin
+            triggerPin.high();
+            TimeUnit.MICROSECONDS.sleep(10);
+            triggerPin.low();
+
+            // Waiting for the echo pin to go high, signaling the start of the pulse return
+            long start = System.nanoTime();
+            while (echoPin.isLow()) {
+                if ((System.nanoTime() - start) / 1000 > 20000) { // 20 ms timeout
+                    log.fine("Timeout waiting for echo signal start");
+                    distance=0;
+                    return;
+                }
+            }
+            long startTime = System.nanoTime();
+
+            // Waiting for the echo pin to go low, signaling the end of the pulse return
+            while (echoPin.isHigh()) {
+                if ((System.nanoTime() - startTime) / 1000 > 30000) { // 30 ms timeout
+                    log.fine("Timeout waiting for echo signal end");
+                    distance = 0;
+                    return;
+                }
+            }
+            long endTime = System.nanoTime();
+
+            // Calculate the distance
+            calculateDistance(endTime - startTime);
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
[…]
