### 2025-04-27 7ac6f93
Add unit tests for RotaryEncoderHelper
