+++ .github/workflows/build-library-jar.yml
@@ -20,11 +20,6 @@ jobs:
         run: rm -rf ~/.gradle/caches
       - name: Change wrapper permissions
         run: chmod +x pi4micronaut-utils/gradlew
-      - name: Setup Gradle
-        uses: gradle/gradle-build-action@v2
-        with:
-          arguments: build
-          build-root-directory: pi4micronaut-utils/
       - name: Execute Gradle build
         working-directory: pi4micronaut-utils
         run: ./gradlew shadowJar
