+++ .github/workflows/build-library-jar.yml
@@ -23,6 +23,9 @@ jobs:
         with:
           arguments: build
           build-root-directory: pi4micronaut-utils/
+      - name: Execute Gradle build with info logging
+        working-directory: pi4micronaut-utils
+        run: ./gradlew build --info
       - name: Execute Gradle build
         working-directory: pi4micronaut-utils
         run: ./gradlew shadowJar
