+++ joss-paper/paper.bib
@@ -43,3 +43,17 @@
     volume  = {18},
     doi     = {10.1002/cae.20217}
 }
+
+@software{RPi.GPIO,
+    author = {Ben Croston},
+    title = {RPi.GPIO: A Python library for Raspberry Pi GPIO control},
+    year = {2024},
+    url = {https://pypi.org/project/RPi.GPIO/}
+}
+
+@software{Diozero,
+    author = {Matt Hawkins},
+    title = {Diozero: Device I/O library for the Raspberry Pi and other platforms},
+    year = {2024},
+    url = {https://github.com/mattjlewis/diozero}
+}
