### 2025-09-02 #7744 @majensen
Thanks @ruthvikm - JOSS prefers the paper code archive to have a name that matches the paper title and author list that matches the paper. Would that be all right?
